Working on Mobile menu, Scrollbar and Education section
Working on:

Mobile menu -  the header curtain needs to be unzipped and opened from right to left;

Scrollbar - modified scrollbar styling;

Education section - All that's missing is the central bar, the bar that appears in the background and the title. Other things may come up.

All titles and subtitles of the body (need corrections)